Extract course cache handling
