+++ syncmymoodle/app.py
@@ -30,6 +30,16 @@ from syncmymoodle.constants import (
     YOUTUBE_LINK_RE,
 )
 from syncmymoodle.context import SyncContext
+from syncmymoodle.course_cache import (
+    cache_root_node,
+    ensure_timemodified_attribute,
+    get_course_cache_root,
+    get_course_node,
+    get_old_node_for,
+    match_old_cache_child,
+    node_from_cache_data,
+    node_to_cache_data,
+)
 from syncmymoodle.filters import (
     as_list,
     configured_patterns,
@@ -51,7 +61,6 @@ from syncmymoodle.storage import (
     load_cookies_from_data,
     read_private_gzip_json,
     save_session_cookies,
-    write_private_gzip_json,
 )
 from syncmymoodle.totp import totp as generate_totp
 
@@ -181,177 +190,28 @@ class SyncMyMoodle:
         self.ctx.downloaded_paths = value
 
     def _match_old_cache_child(self, old_node, child):
-        """Find the previous cache node corresponding to ``child``, if any."""
-        if old_node is None:
-            return None
-        candidates = [
-            c
-            for c in getattr(old_node, "children", [])
-            if c.name == child.name and c.type == child.type
-        ]
-        if not candidates:
-            return None
-        for candidate in candidates:
-            if candidate.url == child.url:
-                return candidate
-        return candidates[0]
+        return match_old_cache_child(old_node, child)
 
     def _node_to_cache_data(self, node: Node, old_node: Node | None = None):
-        timemodified = node.timemodified
-        etag = node.etag
-        is_downloaded = node.is_downloaded
-        # If this file was not (re)downloaded this run but a previously
-        # downloaded version is still on disk, keep the previously cached version
-        # markers. Otherwise the cache would record Moodle's new timemodified/etag
-        # for a file we never actually fetched, which either skips the file
-        # forever or moves the on-disk copy aside as a spurious conflict on the
-        # next run's retry.
-        if (
-            not node.is_downloaded
-            and old_node is not None
-            and getattr(old_node, "is_downloaded", False)
-            and self.get_sanitized_node_path(node).exists()
-        ):
-            timemodified = getattr(old_node, "timemodified", None)
-            etag = getattr(old_node, "etag", None)
-            is_downloaded = True
-        return {
-            "name": node.name,
-            "id": node.id,
-            "type": node.type,
-            "url": node.url,
-            "timemodified": timemodified,
-            "etag": etag,
-            "name_clash_id": node.name_clash_id,
-            "is_downloaded": is_downloaded,
-            "children": [
-                self._node_to_cache_data(
-                    child, self._match_old_cache_child(old_node, child)
-                )
-                for child in node.children
-            ],
-        }
+        return node_to_cache_data(self.ctx, self.invalid_chars, node, old_node)
 
     def _node_from_cache_data(self, data, parent=None):
-        node = Node(
-            data.get("name", ""),
-            data.get("id"),
-            data.get("type", "Unknown"),
-            parent,
-            url=data.get("url"),
-            timemodified=data.get("timemodified"),
-            etag=data.get("etag"),
-            name_clash_id=data.get("name_clash_id", NAME_CLASH_ID_UNSET),
-            is_downloaded=data.get("is_downloaded", False),
-        )
-        node.children = [
-            self._node_from_cache_data(child, node)
-            for child in data.get("children", [])
-            if isinstance(child, dict)
-        ]
-        return node
+        return node_from_cache_data(data, parent)
 
     def cache_root_node(self):
-        """Persist per-course caches into .syncmymoodle_cache files.
-
-        Each course directory beneath basedir receives its own cache file
-        containing the course subtree, which makes caching less brittle than
-        a single global root cache.
-        """
-        if not self.root_node:
-            return
-
-        for semester_node in self.root_node.children:
-            if semester_node.type != "Semester":
-                continue
-            for course_node in semester_node.children:
-                if course_node.type != "Course":
-                    continue
-                course_path = self.get_sanitized_node_path(course_node)
-                # Read the previous course cache before overwriting it, so we can
-                # preserve version markers for files that were not downloaded
-                # this run (see _node_to_cache_data).
-                old_course_root = self._get_course_cache_root(course_node)
-                course_path.mkdir(parents=True, exist_ok=True)
-                cache_path = course_path / ".syncmymoodle_cache"
-                write_private_gzip_json(
-                    cache_path,
-                    {
-                        "format": "syncmymoodle.course-cache.v1",
-                        "course": self._node_to_cache_data(
-                            course_node, old_course_root
-                        ),
-                    },
-                )
+        return cache_root_node(self.ctx, self.invalid_chars, logger)
 
     def _ensure_timemodified_attribute(self, node):
-        # Old cached root nodes might not have the timemodified attribute yet.
-        if not hasattr(node, "timemodified"):
-            node.timemodified = None
-        if not hasattr(node, "etag"):
-            node.etag = None
-        if not hasattr(node, "name_clash_id"):
-            node.name_clash_id = getattr(node, "id", None)
-        for child in getattr(node, "children", []):
-            self._ensure_timemodified_attribute(child)
+        return ensure_timemodified_attribute(node)
 
     def _get_course_node(self, node: Node) -> Node:
-        """Return the enclosing course node for the given node."""
-        cur = node
-        while cur is not None and cur.parent is not None:
-            if cur.type == "Course":
-                return cur
-            cur = cur.parent
-        raise Exception("Node is not part of a course subtree")
+        return get_course_node(node)
 
     def _get_course_cache_root(self, course_node: Node):
-        """Load and return the cached course root for the given course node."""
-        course_path = self.get_sanitized_node_path(course_node)
-        if course_path in self._course_caches:
-            return self._course_caches[course_path]
-
-        cache_path = course_path / ".syncmymoodle_cache"
-        if not cache_path.exists():
-            return None
-
-        payload = read_private_gzip_json(cache_path, "course cache")
-        if not isinstance(payload, dict):
-            return None
-        if payload.get("format") != "syncmymoodle.course-cache.v1":
-            logger.warning("Ignoring unsupported course cache format: %s", cache_path)
-            return None
-        course_data = payload.get("course")
-        if not isinstance(course_data, dict):
-            return None
-
-        cached_course_root = self._node_from_cache_data(course_data)
-        self._ensure_timemodified_attribute(cached_course_root)
-
-        self._course_caches[course_path] = cached_course_root
-        return cached_course_root
+        return get_course_cache_root(self.ctx, self.invalid_chars, course_node, logger)
 
     def _get_old_node_for(self, node: Node):
-        """Return the cached node for this node from the course cache, if any."""
-        try:
-            course_node = self._get_course_node(node)
-        except Exception:
-            return None
-
-        cached_course_root = self._get_course_cache_root(course_node)
-        if cached_course_root is None:
-            return None
-
-        full_path = node.get_path()
-        course_path = course_node.get_path()
-        # Compute the path segments beneath the course root
-        rel_segments = full_path[len(course_path) :]
-        if not rel_segments:
-            return cached_course_root
-
-        try:
-            return cached_course_root.go_to_path(rel_segments)
-        except Exception:
-            return None
+        return get_old_node_for(self.ctx, self.invalid_chars, node, logger)
 
     def _get_or_add_child(self, parent_node, name, id, type):
         for child in parent_node.children:
